backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary>=2.9.9     # ← add this (most reliable & easiest on Render)
pydantic>=2.0
python-dotenv
reportlab
python-multipart==0.0.7
aiosmtplib
python-jose[cryptography]
passlib[bcrypt]
pytest
email-validator
alembic